Simplify langfun by removing MessageTransform.
The concept of MessageTransform adds a flavor of Tensorflow (constructing graphs) to langfun, which is unnecessary complex and is hard to debug. Plus, the effectiveness of structured parsing/prompting minimizes the need for a flexible message transform pipeline. This CL completely removes this concept, so LangFunc could be used as a regular Python function. Consequently, `as_structured` is also removed - users should always use `lf.call` with schema in such cases.

PiperOrigin-RevId: 575556622