Refactor `lf.coding.python.exectuion` and support automatic sandboxing.
In many cases, generated code will include creation of new functions, classes, or return objects that cannot be serialized and communicated across processes. In such scenarios, we need to run the code within the current process. We support a new mode for sandboxing, by specifying `sandbox=None` (which is the default), we will always try run code in a sandbox (separate process) first, if the output could not be serialized/deserialized, we will try rerun the code within the current process - it is safer at this point as the remote access has verified that the code will not run beyond the time limit.

PiperOrigin-RevId: 581396223