Make `lf.eval.Evaluable.id` automatically inferred.
From prior experiments, we found that the `id` field is not very useful and complicates the eval framework's mental model. With this change, the ID of an leaf experiment will be automatically generated based on its class name and hash. As a result, users no longer need to specify `id` for both suites and evaluation tasks.

PiperOrigin-RevId: 620094831